# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FlawFinder AI API
  description: The Ruthless Business & ML Critic - Backend API
  version: 1.0.0
  contact:
    name: FlawFinder AI Team
    email: support@flawfinder.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.flawfinder.ai/api/v1
    description: Production server
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Login user
      tags: [authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /auth/logout:
    post:
      summary: Logout user
      tags: [authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
  /auth/me:
    get:
      summary: Get current user info
      tags: [authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  # Process flows endpoints
...
  /agentops/pipelines:
    post:
      summary: Register a new ML pipeline
      tags: [agentops]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineCreate'
      responses:
        '201':
          description: Pipeline registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineResponse'

    get:
      summary: Get all pipelines for current user
      tags: [agentops]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of pipelines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PipelineResponse'
  /agentops/pipelines/{pipeline_id}:
    get:
      summary: Get a specific pipeline
      tags: [agentops]
      security:
        - BearerAuth: []
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pipeline details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineResponse'
  /agentops/pipelines/{pipeline_id}/monitor:
    post:
      summary: Capture drift metrics for a pipeline
      tags: [agentops]
      security:
        - BearerAuth: []
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriftMetricCreate'
      responses:
        '201':
          description: Drift metric captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftMetricResponse'
  /agentops/pipelines/{pipeline_id}/retrain:
    post:
      summary: Trigger model retraining for a pipeline
      tags: [agentops]
      security:
        - BearerAuth: []
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retraining triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  retraining_id:
                    type: integer
  /agentops/pipelines/{pipeline_id}/status:
    get:
      summary: Get current pipeline health and status
      tags: [agentops]
      security:
        - BearerAuth: []
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pipeline status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatusResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # Authentication schemas
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    UserRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [admin, analyst, engineer, exec]
          default: analyst
    Token:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
        status:
          type: string
        organization_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Flow schemas
...
    PipelineCreate:
      type: object
      required:
        - name
        - model_path
        - metrics_endpoint
      properties:
        name:
          type: string
        description:
          type: string
        model_path:
          type: string
        metrics_endpoint:
          type: string
        metadata:
          type: object
    PipelineResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        model_path:
          type: string
        metrics_endpoint:
          type: string
        status:
          type: string
        metadata:
          type: object
        user_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DriftMetricCreate:
      type: object
      required:
        - drift_type
        - metric_value
        - threshold
      properties:
        drift_type:
          type: string
          enum: [data_drift, concept_drift, label_drift]
        metric_value:
          type: number
        threshold:
          type: number
        metadata:
          type: object
    DriftMetricResponse:
      type: object
      properties:
        id:
          type: integer
        drift_type:
          type: string
        metric_value:
          type: number
        threshold:
          type: number
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
        pipeline_id:
          type: integer
    PipelineStatusResponse:
      type: object
      properties:
        pipeline_id:
          type: integer
        status:
          type: string
        last_retrain:
          type: string
          format: date-time
        latest_metrics:
          type: array
          items:
            type: object
        health_score:
          type: number 